test: Move testing utilities to test target package
Moves `cache_test.go` and associated testdata to the root directory to sit alongside the refactored `g2.GenerateCacheFS` functionality, addressing review comments. test utils extracted to `cache_test_utils.go` as part of the refactor.
