+++ testdata/cache/clean_basic.txtar
