+++ testdata/cache/generate_eclasses.txtar
