+++ testdata/cache/verify_basic.txtar
